fix: fix project form modal bug
